Add files via upload
This project is aimed at building an intelligent system capable of identifying spam emails using natural language processing (NLP) and machine learning. By analyzing the textual content of emails, the system classifies messages as either spam or ham (non-spam). The model is trained using labeled email data and is designed to be efficient, interpretable, and adaptable. The ultimate goal is to create a reliable filter that can be integrated into email services to enhance user safety and communication quality.